refactor: move format duration into parser package
